requirements.txt
Brotli==1.0.6
certifi==2018.10.15
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.1
-e git+http://github.com/preetamherald/django-imgur#egg=django_imgur
django-tastypie==0.14.2
gunicorn==19.9.0
idna==2.7
imgurpython==1.1.7
Pillow==5.3.0
psycopg2==2.7.5
python-dateutil==2.7.3
python-mimeparse==1.6.0
pytz==2018.5
requests==2.20.0
six==1.11.0
urllib3==1.24
whitenoise==4.1
